Simplify GraphicalRecording component
Basically just moved a bunch of stuff into it's own file, to make the
component file a bit easier to work with
